requirements.txt
# Core RL & Simulation
gymnasium==0.29.1
panda-gym==3.0.7
pybullet==3.2.6
stable-baselines3[extra]==2.3.2

# PyTorch (CUDA support auto-detected)
torch>=2.0.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# Video recording
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Development
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
